Moved helpers file into 'piano' folder
