+++ assets/js/piano/helpers/canvasHelpers.js
